Fixed compiler warnings
Former-commit-id: 884715a3eb37bc47887af5f67b353a9ef4537c3e

+++ cocos2d/CCActionInterval.m
@@ -135,7 +135,7 @@
 {
 	CCFiniteTimeAction *prev = [actions objectAtIndex:0];
 	
-	for (int i = 1; i < [actions count]; i++)
+	for (NSUInteger i = 1; i < [actions count]; i++)
 		prev = [self actionOne:prev two:[actions objectAtIndex:i]];
 	
 	return prev;
@@ -364,7 +364,7 @@
 {
 	CCFiniteTimeAction *prev = [actions objectAtIndex:0];
 	
-	for (int i = 1; i < [actions count]; i++)
+	for (NSUInteger i = 1; i < [actions count]; i++)
 		prev = [self actionOne:prev two:[actions objectAtIndex:i]];
 	
 	return prev;
